+++ source/Processing.Infrastructure/Configuration/EventPublishing/IMessageDispatcher.cs
@@ -12,24 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using System.Threading.Tasks;
 using Google.Protobuf;
-using Processing.Infrastructure.Configuration.EventPublishing;
 
-namespace Processing.IntegrationTests.TestDoubles
+namespace Processing.Infrastructure.Configuration.EventPublishing
 {
-    public class MessagePublisherStub : IMessagePublisher
+    /// <summary>
+    /// Protobuf message publisher interface
+    /// </summary>
+    public interface IMessageDispatcher
     {
-        private readonly List<IMessage> _publishedMessages = new();
-
-        public ReadOnlyCollection<IMessage> PublishedMessages => _publishedMessages.AsReadOnly();
-
-        public Task PublishAsync(IMessage integrationEvent)
-        {
-            _publishedMessages.Add(integrationEvent);
-            return Task.CompletedTask;
-        }
+        /// <summary>
+        /// Publish ProtoBuf message
+        /// </summary>
+        /// <param name="integrationEvent"></param>
+        Task DispatchAsync(IMessage integrationEvent);
     }
 }
