+++ source/Processing.IntegrationTests/Application/TestHost.cs
@@ -142,7 +142,7 @@ namespace Processing.IntegrationTests.Application
             _container.ConfigureMoveInProcessTimePolicy(0, 0, TimeOfDay.Create(0, 0, 0));
 
             // Integration event publishing
-            _container.AddEventPublishing(new MessagePublisherStub());
+            _container.AddEventPublishing(new MessageDispatcherStub());
 
             // Business process responders
             _container.Register<IBusinessProcessResultHandler<RequestChangeOfSupplier>, RequestChangeOfSupplierResultHandler>(Lifestyle.Scoped);
